requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
protobuf==4.25.3
psycopg2-binary==2.9.9
PyYAML==6.0.1
